Move cluster logic to status, remove deprecated releases by InUse state
