+++ service/controller/release/resource/status/create.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	appv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/application/v1alpha1"
+	releasev1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/release/v1alpha1"
 	"github.com/giantswarm/microerror"
 	"k8s.io/apimachinery/pkg/labels"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -41,6 +42,39 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		}
 	}
 
+	// Doing this per-release isn't ideal, can we pass this list to each status reconcile somehow?
+	r.logger.LogCtx(ctx, "level", "debug", "message", "searching for running tenant clusters")
+	var tenantClusters []TenantCluster
+	{
+		// var err error
+		tenantClusters, err := r.getCurrentTenantClusters(ctx)
+		if err != nil {
+			r.logger.LogCtx(ctx, "level", "error", "message", fmt.Sprintf("error finding tenant clusters: %s", err))
+		} else {
+			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("found %d running tenant clusters", len(tenantClusters)))
+		}
+	}
+
+	var releaseInUse bool
+	{
+		// Get two sets of just deduplicated versions
+		releaseVersions, operatorVersions := consolidateClusterVersions(tenantClusters)
+		// Check the set of release versions and keep this release if it is used.
+		if releaseVersions[release.Name] {
+			r.logger.Log("level", "debug", "message", fmt.Sprintf("keeping release %s because it is explicitly used", release.Name))
+			releaseInUse = true
+		} else {
+			for _, o := range key.GetProviderOperators() {
+				operatorVersion := getOperatorVersionInRelease(o, release)
+				// Check the set of operator versions and keep this release if its operator version is used.
+				if operatorVersion != "" && operatorVersions[o][operatorVersion] {
+					r.logger.Log("level", "debug", "message", fmt.Sprintf("keeping release %s because a cluster using its operator version (%s) is present", release.Name, operatorVersion))
+					releaseInUse = true
+				}
+			}
+		}
+	}
+
 	var releaseDeployed bool
 	{
 		releaseDeployed = true
@@ -57,6 +91,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("setting status for release %#q in namespace %#q", release.Name, release.Namespace))
 
 		release.Status.Ready = releaseDeployed
+		release.Status.InUse = releaseInUse
 		err := r.k8sClient.CtrlClient().Status().Update(
 			ctx,
 			release,
@@ -69,4 +104,14 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	}
 
 	return nil
+}
+
+// Searches the components in a release for the given operator and returns the version.
+func getOperatorVersionInRelease(operator string, release *releasev1alpha1.Release) string {
+	for _, component := range release.Spec.Components {
+		if component.Name == operator {
+			return component.Version
+		}
+	}
+	return ""
 }
